+++ src/containers/tutor-profile/profile-info/ProfileContainerDesktop.jsx
@@ -20,6 +20,8 @@ const ProfileContainerDesktop = ({
 }) => {
   const { isDesktop } = useBreakpoints()
 
+  const subjectData = userData.mainSubjects.tutor.map((item) => item.name)
+
   return (
     <Box sx={styles.container}>
       <Box sx={styles.avatarContainer}>
