+++ src/containers/tutor-profile/profile-info/ProfileContainerMobile.jsx
@@ -18,6 +18,7 @@ const ProfileContainerMobile = ({
   userData,
   chipItems
 }) => {
+  const subjectData = userData.mainSubjects.tutor.map((item) => item.name)
   return (
     <Box sx={styles.container}>
       <Box sx={styles.wrapperForPhoto}>
