+++ src/containers/tutor-profile/profile-info/ProfileInfo.styles.js
@@ -16,7 +16,8 @@ export const styles = {
     height: '100%',
     minWidth: '150px',
     minHeight: '150px',
-    borderRadius: '15px'
+    borderRadius: '15px',
+    maxWidth: '500px'
   },
   infoWrapper: {
     maxWidth: '552px'
